Refactor: 각각 클래스를 seach에서 독립
