+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/dto/command/event/RemoveGatheringCommand.java
@@ -1,0 +1,11 @@
+package com.sparta.moim.gathering.gathering.application.dto.command.event;
+
+import java.util.List;
+import java.util.UUID;
+
+public record RemoveGatheringCommand(
+    UUID gatheringId,
+    List<String> users,
+    String memberId
+) {
+}
