+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/presentation/contoller/external/MemberController.java
@@ -2,8 +2,8 @@ package com.sparta.moim.gathering.gathering.presentation.contoller.external;
 
 import com.sparta.moim.common.response.ApiResponseData;
 import com.sparta.moim.common.security.CustomUserDetails;
-import com.sparta.moim.gathering.gathering.application.dto.command.SearchGatheringCommand.JoinGatheringCommand;
-import com.sparta.moim.gathering.gathering.application.dto.command.SearchGatheringCommand.LeaveGatheringCommand;
+import com.sparta.moim.gathering.gathering.application.dto.command.event.LeaveGatheringCommand;
+import com.sparta.moim.gathering.gathering.application.dto.command.event.JoinGatheringCommand;
 import com.sparta.moim.gathering.gathering.application.service.MemberService;
 import com.sparta.moim.gathering.gathering.presentation.dto.request.RemoveGatheringRequest;
 import jakarta.validation.Valid;
